+++ examples/commands/basic_command.rb
@@ -7,11 +7,11 @@ require_relative '../events/basic_event'
 class BasicCommand < Evnt::Command
 
   to_validate_params do
-    stop 'Name should be present' if params[:name].nil? || params[:name] == ''
+    err 'Name should be present' if params[:name].nil? || params[:name] == ''
   end
 
   to_validate_logic do
-    stop 'Name should be foobar' if params[:name] != 'foobar'
+    err 'Name should be foobar' if params[:name] != 'foobar'
   end
 
   to_initialize_events do